some code refactoring of ADVI
