+++ fortuna/prob_model/posterior/laplace/laplace_posterior.py
@@ -6,16 +6,14 @@ from typing import Dict, Optional
 import jax.numpy as jnp
 from flax.core import FrozenDict
 from jax import hessian, lax, vjp, devices, jit, pmap
-from jax.lax import psum
 from jax._src.prng import PRNGKeyArray
 from jax.flatten_util import ravel_pytree
-from jax.tree_util import tree_map
 
 from fortuna.data.loader import DataLoader, DeviceDimensionAugmentedDataLoader
 from fortuna.prob_model.fit_config.base import FitConfig
 from fortuna.prob_model.joint.base import Joint
 from fortuna.prob_model.joint.state import JointState
-from fortuna.prob_model.posterior.base import Posterior
+from fortuna.prob_model.posterior.gaussian_posterior import GaussianPosterior
 from fortuna.prob_model.posterior.laplace import LAPLACE_NAME
 from fortuna.prob_model.posterior.laplace.laplace_approximator import \
     LaplacePosteriorApproximator
@@ -29,11 +27,10 @@ from fortuna.prob_model.posterior.posterior_state_repository import \
 from fortuna.prob_model.prior.gaussian import (DiagonalGaussianPrior,
                                                IsotropicGaussianPrior)
 from fortuna.typing import CalibMutable, CalibParams, Mutable, Params, Status
-from fortuna.utils.nested_dicts import nested_get, nested_set, nested_unpair
-from fortuna.utils.random import generate_random_normal_like_tree
+from fortuna.utils.nested_dicts import nested_get, nested_set
 
 
-class LaplacePosterior(Posterior):
+class LaplacePosterior(GaussianPosterior):
     def __init__(
         self,
         joint: Joint,
