+++ fortuna/prob_model/posterior/normalizing_flow/advi/advi_approximator.py
@@ -33,10 +33,10 @@ class ADVIPosteriorApproximator(PosteriorApproximator):
         self.n_loss_samples = n_loss_samples
 
         if which_params:
-            if type(which_params) != tuple:
+            if not isinstance(which_params, tuple):
                 raise ValueError("`which_params` must be a tuple of lists.")
             for list_keys in which_params:
-                if type(list_keys) != list:
+                if not isinstance(list_keys, list):
                     raise ValueError("Each element in `which_params` must be a list.")
         self.which_params = which_params
